requirements.txt
numpy
python-dotenv
# For web app + API
requests
supabase
flask
# For ML
jupyter
polars[pandas]
matplotlib
catboost
scikit-learn
peewee